requirements.txt
anthropic>=0.20.0
streamlit>=1.36.0
pillow>=10.1.0
pyautogui>=0.9.54
pygetwindow>=0.0.9
python-dotenv>=1.0.0
typing-extensions>=4.8.0 